2157 20230701 = release-0.0.4

+++ ReMarket-02-08/core-service/src/main/java/ru/rrenat358/core/entities/OrderItem.java
@@ -8,6 +8,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -36,10 +37,10 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(name = "price_per_product")
-    private Integer pricePerProduct;
+    private BigDecimal pricePerProduct;
 
     @Column(name = "price")
-    private Integer price;
+    private BigDecimal price;
